requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
google-generativeai==0.3.2
python-socketio==5.10.0
asyncio==3.4.3
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0